One possible solution to group randomizing goal
We use a very basic algorithm to sort a randomized list of numbers from
highest to smallest.

Input arrays:

- original list of student names
- list of numbers, randomly generated, one random number tied to each
student name

Output:

Two new arrays:

- the list of random numbers, now sorted from largest to smallest
- the list of names, that have been added to a new array in the same
order as the sorted numbers.

This provides us with an easy to understand, but not very efficient,
way to shuffle the original list of student names.